[storage-interface] move ExecutedTrees to storage-interface
