+++ execution/executor/src/components/block_tree/test.rs
@@ -5,8 +5,9 @@ use crate::components::block_tree::{epoch_genesis_block_id, BlockLookup, BlockTr
 use aptos_crypto::{hash::PRE_GENESIS_BLOCK_ID, HashValue};
 use aptos_infallible::Mutex;
 use aptos_types::{block_info::BlockInfo, epoch_state::EpochState, ledger_info::LedgerInfo};
-use executor_types::{ExecutedChunk, ExecutedTrees};
+use executor_types::ExecutedChunk;
 use std::sync::Arc;
+use storage_interface::ExecutedTrees;
 
 impl BlockTree {
     pub fn new_empty() -> Self {
