+++ execution/executor/src/components/chunk_commit_queue.rs
@@ -6,9 +6,9 @@
 use anyhow::{anyhow, Result};
 
 use crate::components::in_memory_state_calculator::IntoLedgerView;
-use executor_types::{ExecutedChunk, ExecutedTrees};
+use executor_types::ExecutedChunk;
 use std::{collections::VecDeque, sync::Arc};
-use storage_interface::DbReader;
+use storage_interface::{DbReader, ExecutedTrees};
 
 pub struct ChunkCommitQueue {
     persisted_view: ExecutedTrees,
